+++ lib/src/lmi_oracle.cpp
@@ -15,7 +15,7 @@ std::optional<Cut> lmi_oracle::operator()(const Arr& x)
 {
     const auto n = x.size();
 
-    auto getA = [&, this](unsigned i, unsigned j) -> double {
+    auto getA = [&, this](size_t i, size_t j) -> double {
         auto a = this->_F0(i, j);
         for (size_t k = 0U; k != n; ++k)
         {
